[mbox] Ignore messages without 'Message-ID' or 'Date' headers
These headers are mandatory in any message, so those messages
without them are ignored.

In fact, the 'mbox' class from the 'mailbox' module produces
invalid messages when a line on a message body starts with
a 'From ' string. In those cases, an invalid message is generated
using the remaining part of the body. Checking if the message
has some headers (like 'Message-ID' or 'Date') avoids the problem,
ignoring them.