[backend] Change item structure
Items generated using the metadata decorator will change their
structure with this patch.

Metadata is no longer stored under '__metadata__' key. Instead
of this, metadata will be stored at the top level of the item
while the contents of the item will be stored under 'data' keyword.

This is an example of the new item structure:

  {
   'backend_version' : '0.1',
   'backend_name' : 'mock backend',
   'perceval_version' : '0.0.1',
   'timestamp' : 1459334803.0,
   'origin' : 'example.com',
   'uuid' : '0caaf24ab1a0c33440c06afe99df986365b0781f',
   'updated_on' : 1459334803.0,
   'data' : {
             'id' : 8,
             'title' : 'Mock item',
             'description' : 'A short description of this item'
            }
  }

Specific unit tests for each backen were also updated.