[github] Fix error using utcnow() and timestamp() in chain
When timestamp() method is called, it applies the timezone
of the system where it is called. If utcnow() is called
before, the result date will not be in UTC, it will be in the
timezone added by timestamp().

To fix this problem, use time.time() that does the same
that calling to datetime.datetime.utcnow().timestamp() .