[gmane] Add Gmane backend class
On this backend, the method fetch does not supports to filter
by 'from_date' like in other backends because this is not supported
on Gmane. Instead, 'offset' filter is available.

This commit also includes a new decorator for the fetch method of
this class. This decorator adds the field 'offset' to the metadata
section, needed for pagination and resume.