[telegram] Include 'offset' metadata on each item
Due to Telegram backend only allows to fetch items passing an
'offset' value and not the typical 'from_date' parameter,
that it is used in most of the backends, it is usefull to include
the offset of each item as metadata.

This commit adds the decorator 'telegram_metadata()' to include
this keyword in metadata section.

This commit updates Telegram backend to 0.2.0 version.