[stackexchange] Sleep the client when 'backoff' field is received
According to StackExchange documentation, "a dynamic throttle
is also in place on a per-method level. If an application
receives a response with the backoff field set, it must wait
that many seconds before hitting the same method again."

This patch makes the client to wait the seconds set on 'backoff'
field before sending a new requests.