[tests] Fix invalid mock server config in stackexchange test
On 'test_get_questions_more()' test, the httpretty configuration
for the mock HTTP server was wrong. It worked because httpretty
push the bodies that will be returned to a queue, but the coded
behaviour was wrong. This patch fixes that problem.

The test has been renamed to 'test_get_questions_pagination'.