[mediawiki] Add test flag to ignore MAX_RECENT_DAYS exception
On Mediawiki pre-1.27 servers, due to server restrictions,
'from_date' parameter only works for dates not older than
MAX_RECENT_DAYS. Usually, this parameter is configured to
30 days. Due to this, unit tests started to fail when the
date that was set was 30 days than the current date.

To fix this problem, a flag has been added to the backend to
ignore this date check.