[phabricator] Fetch data about authors that are not users
Actores such as bots, scheduled tasks, etc, can be found as
authors (normally in transactions) in Phabricator tasks.
Using the 'users' API methods to get their data returns an empty
result and made chashing the code. To fetch these data, it is
needed to use to 'phids' API method.

Backend version incresed to '0.1.1'.