[pipermail] Fix error storing decompressed mbox files
By default, requests library decompresses gzip-encoded
responses, but in some cases is only able to decompress
some parts due to encoding issues or to mixed contents.

This patch fixes this problem storing the orinal/raw data
(compressed or decompressed) into the mbox file to process
it in the next phase using the decoder of the MBox class.

Backend version bumped to '0.4.1'.