[tests] Make sure tests import the right backend classes
Since the introduction of the namespace for backends, tests
first try to import the packages installed in the system
by setuptools, ignoring the classes from this repository.

This hack fixes the problem updating the sys.path and
declaring the 'perceval.backends'. This declaration will
make packages to point to development packages insted
of those installed in the system.