[askbot] Fetch full list of comments
Before, by parsing the HTML we were unable to retrieve the full list of
comments inside each element. Now we do a different request to fetch the
comments by a given object_id. The content displayed in the item will be
exactly the same as the returned one.

Backend version updated to 0.2.0