[bugzilla] Set maximum number of bugs requested on CSV queries
The 'buglist' method fetches a CSV stream  that contains a
summary of bugs. By default, it fetches 10000 bugs but in
some busy servers the request lasts enough to be disconnected
before fetching the list.

A new parameter 'max_bugs_csv' has been defined to avoid these
problems limiting the number of bugs requested on each CVS
query.

Backend version updated to '0.6.0'.
This mission of this task is to include a parameter to limit the number of bugs download on each call.