[supybot] Generate item ids using the body of the message
Two or more IRC messages can be sent by the same user at the same
time. As the identifier of an item is generated using the date
and the nick of the sender, it causes that two different messages
could have the same identifier.

This commit fixes this bug using the body of the message to
generate the identifier of each item.

Backend version bumped to 0.5.0.