[git] Fix error fetching items from an empty repository
On empty repositories, those which do not have any history or
are only initialized, there are some commands that cannot
be run, like a pull or log. If any of this commands is called
an error will be returned. Thus, fetch() method fails when
a repository is empty.

This commit fixes this problem, generating an empty iterator
for those calls on an empty repository.

Backend version bumped to '0.6.1'.

Fixes #102