[mbox] Fix decoding error on unix-from headers
The mbox class from Python's mailbox module fails when it tries
to decode non-ascii unix-from headers. This header is used as
a separator between messages. When this error is found, the
class stops reading messages from the mbox.

To fix this bug it was necessary to wrap mbox class to override
the way messages are readed from the mbox file, caching the
exception and decoding using UTF-8.

Backend version increased to 0.7.2.