[github] Manage 404 not found error when getting the public github orgs for a login.
Sometimes GitHub API when calling  https://api.github.com/users/joltup/orgs return
404 Client Error: Not Found for url: https://api.github.com/users/joltup/orgs. But
if you try to call manually this url, it works and returns an empty array in
the cases detected.

This is an error that must be managed inside the backend because the user of the
perceval backend can not manage it and continue with the rest of the logic for
extracting a raw github item.

This orgs information is not critical and it is filled with an empty array if the
404 not found exception is raised, the normal response in these cases from the
GitHub API when it does not fail.