[git] Run Git commands setting HOME environment variable
The $HOME env variable is where Git looks for the configuration
file. So far, Git commands were run without setting this
variable. This patch fixes this error and allows to use
a configuration file when Git is run.

More info about the environment variables used by Git can be
found on the next link:

  https://git-scm.com/book/gr/v2/Git-Internals-Environment-Variables

Backend updated to 0.8.1.

Fixes #149