[git] Fix fetch-pack for HTTPs connections
The command 'fetch-pack' only works with the TCP (git://)
protocol unless '--stateless-rpc' flag is active, the caller
has established a connection with the server using HTTP
and a pipe to it.

Some Git servers do not let to talk to them using TCP
protocol, only using HTTP(s) on top of it.

To fix this issue, dulwich library has been used. It allows
to talk to Git repositories without calling git commands
directly. In this particular case, it is able to fetch
packs using any protocol.

This would be the first step to replace Git commands by
calls to a pure Python library.