[git] Limit the number of 'have' commits sent by fetch-pack
The current code sends, during the fetch-pack negotiation,
the hash for all the commits that the client has. This is not
needed. Only head hashes are usually needed by the server
to determine which objects needs to send back to the client.
In big repositories, this may hang the communication.

This commits adds a simple graph walker which only sends
the current hashes for each 'refs/heads/' or 'refs/tags/'
available in the client repository.

Backend version updated to 0.8.5.