[git] Fix invalid URL when TCPGitClient is initialized
When 'git://' URLs were in use, the initialization of TCPGitClient
instances were wrong because the host given to it was invalid.
When an instance tryed to connect to a Git host, the code raised
a "Name of service not known" error.

To fix this bug, we delegate to dulwich the initialization
of clients according to the protocol that will be used.

Backend version updated to 0.8.7.