[backend] Fix 'find_backends' to get all backend descendants
This function uses the class method `__subclasses__` to get
derived classes from Backend and BackendCommand. The problem
with this method is it only returns direct/children subclasses,
not grandchildren classes.

MBox is a children of Backend class while Pipermail and Hyperkitty
are children of MBox and grandchildren of Backend. So far, only
MBox was found because it's a direct subclass of Backend.