[git] Request in fetch-pack the outdated refs only
The current code, during the fetch-pack negotiation, sent the
hashes of all the references published by the server. This is
not needed as the client only needs the outdated references
to keep the repository in sync.

This commit reduces the number of 'want' references, sending
the outdated ones only.

Backend version updated to 0.8.9.