[meetup] Remove 'draft' status due to a bug in Meetup API
This is a work around to fix the problem with Meetup API.
In some cases, having this status in the list of requested
events, makes the API to return zero events.

The bug was reported in https://github.com/meetup/api/issues/260.
As soon as it gets fixed, we will add it back again to the list
of requested events.

Backend version updated to 0.7.0.