[git] Fix cloned directory of local repository
This patch fixes the issue #262. The git_path is derived by
joining the default Perceval location plus the path of the local
Git repository to analyse, stripped of the initial slash. The
initial slash in a component of a os.path.join causes all
previous components to be thrown away and the joining
continues from the absolute path component.