[git] Parse additional pattern for moved files
There was a pattern for moved files in Git that was not captured
by Perceval. When the destination directory and filename are both
different, the pattern in the stats part of the Git log is like
the following:

    dir1/file1.txt => dir2/file2.txt

In the files dictionary, Perceval created three items (instead of
two). The additional file had "dir1/file1.txt => dir2/file2.txt"
as name.

This patch fixes this problem, by adding an additional pattern
for moved files to the stats parsing.