[archive] Add Archive class
The data returned during fetch executions would be stored in
different archives (one per execution) for further recovering.
It will allow to transparently play back previous fetch executions,
enabling the recovery of the stored data as it was originally
collected from the repository.

This new system will replace the current cache system, which its
functionality is similar to the archive.