[mbox] Define 'tmp_path' variable before copying mboxes
This variable is not created when an error occurs copying mboxes.
It makes the code fail because the finally block that handles
OSError exceptions requires that variable.

Aditionally, this commit improves the test where this copything
mboxes functionality is tested. So far, the test would fail when
super users run it.

Backend version updated to 0.9.1