[backend] Remove empty category from parsed arguments
The 'category' argument does not have a default value. When
parameters are parsed and a category is not given, its value
is set to 'None'. This caused an error due to backends
do not consider this value as a valid category.

This patch removes the category from the parsed parameters
when the value is empty or None.