[graal] Add Generic Repository AnALyser backend
This class inherits from Git backend, thus it fetches the commits
from a local Git repository and enables to add the result of
the analysis within the `analysis` attribute of the Perceval
item returned. To initialize this class, you have to provide
the local path of a Git repository (URI), a value for
`git_path`, where the repository will be mirrored, and the path where
a working tree will be created. The working tree is added to the
mirror and removed after the analysis is over.

For each target commit (by default all of them), a checkout version
of the repository is created at `worktreepath` to ease the analysis.
Note that you can customize the filter to select commits, by
redefining the method `_filter_commit(self, commit, files)`.
Furthermore, you can plug your analysis by redefining the
method `_analyse(self, commit)` as well as tweak the item
generated by redefining the method `_post(commit)`.