[graal] Add Generic Repository AnALyser backend
This code includes Graal, a generic backend to support code
analysis. It extends the Git backend, thus it fetches the
commits from a local Git repository, and enables to add
the result of the analysis within the `analysis` attribute of
the Perceval item returned. To initialize this class, you have
to provide the URI repository and a value for `git_path`. For
each target commit (by default all of them), a snapshot of the
repository can be created to ease the analysis. The parameter
`snapshot` allows to manage the snapshot as a tar object file,
tar file or folder.
Note that you can customize the filter to select commits, by
redefining the method `_filter_commit(self, commit, files)`.
Furthermore, you can plug your analysis, by redefining the method
`_analyse(self, hash, snapshot)` as well as tweak the item
generated by redefining the method `_post(commit)`.