Fix support for package namespaces.
This module was not really using package namespaces. This
patch should provide that support. For that, some changes
had to be done to the module itself, mainly changing
__init__.py in perceval and perceval/backends to contain
only the line for declaring namepaces. This means that
symbols that were exported by those __init__.py files
now need to be imported from the corresponding modules.
In particular, __version__ is now provided by
perceval.backends.core for this package.

All of this has bin fixed in bin/perceval, but maybe
other packages (GrimoireELK, SirArthur, Mordred)
would need similar fixes.