[mbox] Allow support for zip file in MboxArchive
This code enhances the `MboxArchive` class with zip support, thus
enabling the analysis of mbox archives compressed with zip. Note
that only the first compressed file within a zip is set as
container of the MboxArchive. In case the zip contains more than
one file, an error is included to the log.