[groupsio] Fetch archives from group-specific domains
This code allows the groupsio backend to fetch archives of
organizations having their own domains (e.g., lists.cloudfoundry.org).
Given a group name and a token, the corresponding group id is
searched among the list of subscriptions of the token. If the
group id is not found, an error is thrown.

Groupsio test code has been changed accordingly.