[github] Handle upstream server connection problems
This code limits the connection problems with the github api. In a
nutshell, the github backend puts now in place retrying mechanisms for
the following http errors 403, 500, 502, 503. Note that this solution
doesn't guarantee to completely solve the connection problems, since
the retry is executed a limited number of times (param `max_retries`),
before raising an exception.