[github] Fetch repo information
This code allows to fetch repo information of a target repository. The time when the
repo was fetched (i.e., `fetched_on`) is included in the JSON document produced
by Perceval, and it is used for generating the uuid and defining the updated
on and id attributes of the document. The category to execute this
backend is `repository` and an example of command is:
```
perceval github kubernetes kubernetes
--category repository -t ... --sleep-for-rate
```