[github] Add 'repository' category to fetch metadata from repositories
This code allows to fetch repositories information of a target
repository. The time when the repo was fetched (i.e., `fetched_on`)
is included in the JSON document produced by Perceval, and
it is used for generating the uuid and defining the updated
on and id attributes of the document.

The category to execute this backend is `repository`.