[confluence] Extend Perceval Confluence backend to index ancestor pages
The Perceval Confluence backend can parse the entire tree of
Confluence pages, indexing each page as separate document;
this way, the information re. page hierarchy gets lost
(ie, is this page a child of this other page?)

In order to allow understanding if a page (represented by a document
in Perceval) is a (direct or indirect) child of a page (given it's
Web URL), we have added a new document field called 'ancestors',
which contains a list of URLs that represent all the pages that are
a parent of the current one.

Example: Take the following pages:
    - https://finosfoundation.atlassian.net/wiki/spaces/FDX/pages/75530554/Code+Hosting , child of ...
    - https://finosfoundation.atlassian.net/wiki/spaces/FDX/pages/75530388/Development+Infrastructure , child of ...
    - https://finosfoundation.atlassian.net/wiki/spaces/FDX/pages/75530320/The+Open+Developer+Platform+ODP

The Perceval document of the 'Code Hosting' page will now include a
new 'ancestors' field with value '['spaces/FDX/pages/75530388/Development+Infrastructure',
'spaces/FDX/pages/75530320/The+Open+Developer+Platform+ODP']'

Resolves #486