[github] Handle pull request reviewers in Enterprise API
The public API returns a dict from '.../pulls/{number}/requested_reviewers'
with users and teams keys. Github Enterprise returns a plain
list of users.

This code fixes #523 and allows to handle the second case, thus
the plain list of users is converted to a dict with the following
structure: `{'users': group_requested_reviewers}`.

Signed-off-by: Cedric Williams <cewilliams@paypal.com>
Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>