[gitlab] Force UTF-8 encoding on response contents
The module 'requests' sometimes tries to guess the encoding
of the responses received. It uses the module 'chardet'
to determine the encoding. In the case of GitLab it is unable
to determine the encoding and usually produces bad guesses
as Big5 encoding when all it is encoded as UTF-8.

As we know GitLab encodes in UTF-8, this commit forces
to parse the response using that encoding.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>