[gitlab] Recalculate outdated list of merge requests
If during the fetching process, a merge request was modified, the
initial list of merge requests is outdated. Due to it, returned
items will not be ordered, which can cause several problems when
an incremental retrieval is performed.

To avoid this problem, when the backend detects that an MR has been
updated, the process will not be cancelled but it will request a new
list, starting on the 'updated_on' field of the last item returned.

Backend version updated to 0.8.0.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>