[backend] Add search fields in metadata items
This code enhances the metadata information with
a set of search fields, which simplify query operations
and avoid the manual inspections of the Perceval items.

These fields are added to the item metadata information
in the `search_fields` attribute. By default, `search_fields`
contains the id of the item ('item_id': item_id_value), obtained
via the `metadata_id` method). However each backend can set
extra search fields using the dict EXTRA_SEARCH_FIELDS. An
example of EXTRA_SEARCH_FIELDS is provided below:

```
   {
      'project_id': ['fields', 'project', 'id'],
      'project_key': ['fields', 'project', 'key'],
      'project_name': ['fields', 'project', 'name']
   }
```

Each key in the dict is a search field to be included in the item
metadata information, while the corresponding value is a list that
stores the "path" of the search field value within the item.

Tests have been added accordingly.
The backend version is set to 0.9.0.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>