[backend] Add search terms in metadata items
This code enhances the metadata information with
a set of search terms, which simplify query operations
and avoid the manual inspections of the Perceval items.
The search terms are included in a dict with the
following shape:
        {
            'term-1': 'value-1',
            'term-2': 'value-2',
            'term-3': 'value-3',
        }

The search terms are added to the metadata information
of each item in `search_terms` attributes. If `search_terms`
is not set, it will be set to `None`.

Tests have been added accordingly.
The backend version is set to 0.9.0.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>