[perceval] Add summary class
This code includes a class that models the summary of a
fetch execution. It includes information such as the
number of items fetched, the min and max dates of the
items fetched, the last UUID generated and an extra
field that can be modified by each backend.

Tests have been added accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>