[github] Add search fields
This code extends the GitHub backend by including
a set of search fields to simplify query operations.
For the categories `issue` and `pull_request`, the
search fields include the issue/pull request number,
labels, state and the name of the milestone. For the
category `repository`, license and language are set
as search fields. All categories include the `item_id`
as search field.

Tests have been added accordingly.
The backend version is set to 0.23.0.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>