[meetup] Pass oauth2 token via headers
The oauth2 token was passed as payload of the requests when
issuing petitions to the Meetup API. This was causing a potential
security leak for failing petitions, since the URL with the
corresponding payload was fully exposed.

Now the token is passed through the headers, which are then
sanitized before storing them to the archive.

Since API tokens have been decommissioned in favor of
oauth2 token, API token support has been removed from
the backend.

Tests have been changed accordingly.
Backend version is now 0.15.0.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>