[slack] Pass API token via headers
The API token was passed in the payload of the requests when
issuing petitions to the Slack API. This was causing a potential
security leak for failing petitions, since the URL with the
corresponding payload was fully exposed.

Now the token is passed through the headers, which are then
sanitized before storing them to the archive.

Tests have been changed accordingly.
Backend version is now 0.7.4.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>