[backend] Organize BackendItemsGenerator parameters
The parameters defined in BackendItemsGenerator methods
do not follow an order. In some methods archive parameters
go first and in others at the end, common parameters are
in between archive ones and so on.

With this commit, methods are align following a common
approach, where backend arguments go first followed
by common and filtering ones. Archive paramters go at
the end of the methods.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>