[tests] Generate different 'updated_on' values for backend tests
So far, the same value of 'updated_on' was assigned for
all the mocked items generated. This does not allow to tests
whether the Summary class is assigning correcly min, max and
last attributes for 'updated_on' since all are the same.

With this patch, these values will be different for each item.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>