[backend] Add support for blacklisting items
This code extends Perceval to support the
blacklisting of remote items. In a nutshell,
the Namedtuple OriginUniqueField is defined
to model the origin unique field (composed of a
`name` and `type`).
The OriginUniqueField is defined by each backend
and used by the method `_skip_item` to filter
out the blacklisted items.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>