[github] Handle `None` requested reviewers
This code handles `None` values for requested
reviewers. Before accessing the login attribute, a
condition checks that the reviewer is not None and
the `login` attribute exists.

An example of this behaviour is visible at:
hyperledger/sawtooth-core#93, where the last requested
reviewer (on the right side) is None, and only a yellow
circle (without name) is shown on the web page.

Backend version is now 0.24.1
Tests have been added accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>