[groupsio] Fetch messages after a given date
This code allows to collect a list of messages sent after
a given date (included) from the groupsio API. The backend
`from_date` parameter is propagated to the client, which sends
it to the API endpoint to download archives. This change is
made possible after a recent change in the upstream API, and
allows to improve the performance of the fetching operation.

Tests have been added accordingly.

Backend version is now 0.3.2

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>