[github] Fetch issue/pull request events
This code adds a new category to fetch the
following event types from issues and pull
requests:
- ADDED_TO_PROJECT_EVENT
- MOVED_COLUMNS_IN_PROJECT_EVENT
- REMOVED_FROM_PROJECT_EVENT
- CROSS_REFERENCED_EVENT
- LABELED_EVENT
- UNLABELED_EVENT
- CLOSED_EVENT

Due to the limitation of not fetching events
after a given date from the endpoint `timeline`
of GitHub v3, the events are fetched via
the GitHub v4 (based on GraphQL).

The events are fetched in the following way. The
issues of a given tracker are retrieved in
ascending order based on the last time they
were updated. For each issue, its events (optionally
from/to a given date) are collected using a GraphQL call.
Each event is returned by Perceval together with
the corresponding issue (available in `data.issue`).

No user information beyond the `login` is included in
data returned by Perceval for this category. Thus,
the filter classified support has no effect when
fetching events.

The backend can be executed in the following way:
```
perceval github chaoss grimoirelab-toolkit
--category event -t xxx
--sleep-for-rate
--no-archive
--from-date 2019-01-01
```

Tests have been added accordingly.
Backend version is now 0.26.0

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>