[perceval] Add githubql backend
This class allows the fetch the issue events of a GitHub
repository. Note that the events retrieved included also the
ones of pull requests, since in GitHub, every pull request
is an issue, but an issue may not be a pull request. Pull
requests can be identified by the attribute `pull_request`
included in `data.issue`.

The following event types from issues and pull
requests:
- ADDED_TO_PROJECT_EVENT
- MOVED_COLUMNS_IN_PROJECT_EVENT
- REMOVED_FROM_PROJECT_EVENT
- CROSS_REFERENCED_EVENT
- LABELED_EVENT
- UNLABELED_EVENT
- CLOSED_EVENT

Due to the limitation of not fetching events after a
given date from the endpoint `timeline` of GitHub v3, the
events are fetched via the GitHub v4 (based on GraphQL).

All issues of a given tracker are retrieved in ascending order
based on the last time they were updated. For each issue, its
events (optionally from/until a given date) are collected using
a GraphQL call. Each event is returned by Perceval together
with the corresponding issue (available in data.issue).

Since the events are collected issue by issue, the incremental
fetching is not supported. This limitation is due to the fact
that events that occur on an issue may not update the issue
attributes. Since there is no way to identify new events from
the attributes of an issue, all issues must be fetched for
every execution.

No user information beyond the login is included in data
returned by this backend. Thus, the backend doesn't require
filter classified support.

The backend can be executed in the following way:
```
perceval githubql chaoss grimoirelab-toolkit
--category event -t xxx
--sleep-for-rate
--no-archive
--from-date 2019-01-01
```

Tests have been added accordingly.
Backend version is 0.1.0

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>