[slack] Update resource endpoints
This commit updates the resource endpoints
to fetch the messages and information of a
channel. The endpoints changed are as follows:
* channels.history -> conversations.history
* channels.info -> conversations.info

This change is required since the `channels.*`
menthods are deprecated since Jan 7, 2020.
Slack suggests the usage of 'conversations.*'
methods as the alternative.
Ref: https://api.slack.com/changelog/2020-01-deprecating-antecedents-to-the-conversations-api

The tests have been updated accordingly.
The backend version is now 0.9.1

Signed-off-by: Animesh Kumar <animuz111@gmail.com>