[doc] Add Slack API token generation info
This commit adds documentation for creation of
API token for the Slack backend.
Slack apps can be used to generate an API token.
A script to generate the token is also linked.

This change was necessary since the documentation
was lacking this information. Additionally, legacy
tokens are deprecated since May 5, 2020. Slack
proposes the use of Slack apps as the new alternative.
Ref: https://api.slack.com/changelog/2020-02-legacy-test-token-creation-to-retire

Signed-off-by: Animesh Kumar <animuz111@gmail.com>