[slack] Handle messages without user/bot ID
The user/bot ID is used by the backend to create
the the Perceval document ID. In the scenario where
this information is missing (e.g., bot deleted), the
document ID cannot be generated and error is thrown
in the method `metadata_id`.

This code proposes to set the username as fallback
value for the messages without user/bot ID, thus
allowing the creation of the Perceval document ID.

Tests have been added accordingly.
Backend version is 0.9.2.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>