[weblate] Initial import
This code proposes an initial version of the backend to extract
changes from a Weblate instance. The changes can be collected
for a given project or for all the projects in the instance.

The initial tests have been done on:
- https://hosted.weblate.org (Weblate API 4.2)
- https://translations.documentfoundation.org (Weblate API 4.1)

Notes:
- Sleep for rate is working in >=4.1.
- Incremental fetching works for all projects in >=4.1.
- Incremental fetching works for single projects in >=4.2.
- Users details can be fetched only if the token has enough permissions.
  See details at: https://docs.weblate.org/en/latest/api.html#get--api-users-.

The backend can be executed as follows:
```
perceval weblate
https://translations.documentfoundation.org
-t
<token>
--from-date
2020-01-01
--no-archive
```

The token can be obtained after registering to a weblate
instance (e.g., https://translations.documentfoundation.org/),
via the page <instance>/accounts/profile/#api

Signed-off-by: Quan Zhou <quan@bitergia.com>