[travis] Support for python 3.6, 3.7 and 3.8
This code aims at aligning the CI tests across the
different grimoirelab components.

Python 3.5 is removed as its support period has ended and
Python 3.6, 3.7 and 3.8 are now supported.

The version of setuptools and pip has been downgraded as a
hotfix to solve the failing CI tests. It must be related to
the issue https://github.com/pypa/setuptools/issues/2353.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>