[docs] Expand docs for implementing new backends
Using the current documentation, it can be pretty difficult to learn how
to appropriately implement a new backend for perceval.  Many of the
methods that backends are supposed to implement have no documentation
explaining what the methods are supposed to do, or even how their
arguments are formatted.

Although it is possible to reverse engineer this information from other
backends, this process is time consuming, frustrating, and may deter new
developers from being able to use perceval with their application.

This commit expands documentation relating to implementing new backends
in several ways:

- The class-level documentation for backends has been expanded to list
more of the necessary methods and attributes that backends are
responsible for implementing, and additionally lists a brief summary of
each method.
- Several methods which were previously not listed (_init_client), or listed as
optional (metadata_*) have been listed as mandatory.  As far as I can
tell, there is no way to create a working backend without implementing
these methods, or at least stubbing them.
- Information about class attributes in the class-level documentation
has been replaced with a brief summary and a link, with the full text
being moved to the attribute itself, making the class-level
documentation easier to digest.
- Some of the documentation for class attributes that was sparse or
missing altogether has been added
- All* methods and class attributes that are intended to be implemented
by subclasses of `Backend` have detailed documentation describing
exactly what is expected of the implementor, and giving context for some
of the terms or argument names used where necessary.
- Some references to other methods and class attributes in the docs have
been converted to links, to make navigating the docs easier.

Some important notes:

- The documentation for `has_resuming` has been left blank.  Despite my
efforts, I wasn't able to determine what this was meant to indicate, and
could find no examples of it being used in context with which to
reverse engineer its purpose.  If someone is able to fill me in on what
this method does, I would be happy to add documentation in.
- I haven't been working with perceval very long, and most of the
knowledge I am contributing here is based on experimentation, reverse
engineering, my own experiences working on perceval-osf, and
extrapolating from what documentation existed before hand.  It's very
possible that I've made a mistake in describing some of these methods,
so please review carefully to make sure I am correct.  If I've made any
mistakes, I will be happy to be corrected, and revise my mistake quickly
after.

If you have any other advice, or recommended changes, I would be happy
to hear!

Addendum 1 (fa2cc41): Apply @sduenas's style format corrections

See https://github.com/chaoss/grimoirelab-perceval/pull/738#pullrequestreview-693785417

Co-authored-by: Santiago Dueñas <sduenas@bitergia.com>

Addendum 2 (70b0c4e): Add documentation for `has_resuming`

This is based primarily off of the PR comment written by
@sduenas@github.com linked below.  Most of the information included here
is based off of this comment, and not off of reverse engineering, actual
knowledge, or anything else, so please review before accepting.

https://github.com/chaoss/grimoirelab-perceval/pull/738#issuecomment-869604429

Addendum 3: Correct final double space

https://github.com/chaoss/grimoirelab-perceval/pull/738/files#r677386098

Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 2006 The Linux Foundation and its contributors.
1 Letterman Drive
Suite D4700
San Francisco, CA, 94129

Everyone is permitted to copy and distribute verbatim copies of this
license document, but changing it is not allowed.

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.

Signed-off-by: Emi Simpson <emi@alchemi.dev>