[mattermost] Change the key from 'data.metadata.images' to value
This code converts the key (URL) to value when the field
'data.metadata.images' exists, thus the ES index does not have a
field for each image URL and consequently avoid the error
'Limit of total fields [1000] in index [...] has been exceeded'.

From:
```
{
  ...
  "images": {
    "https://.../api/v4/image?...": {
      "width": 1280,
      "height": 640,
      ...
    }
  }
}
```

To:
```
{
  ...
  "images": [
    {
      "url": "https://.../api/v4/image?...",
      "width": 1280,
      "height": 640,
      ...
    }
  ]
}

```

Test updated accordingly.

Version updated to 0.5.0.

Signed-off-by: Quan Zhou <quan@bitergia.com>