[mattermost] Support running with channel names
The current mattermost backend runs using a channel ID in order to
select which channel is being scraped.  However, this can get in the way
in certain circumstances.  Mainly, the channel ID can't be derived from
a mattermost URL like

https://my-mattermost.link/some-team/channels/some-channel

This PR allows the mattermost backend to be called with a team name and
a channel name as an alternative to the normal behavior.

This means that tools like grimoirelab-elk would be able to accept a
normal mattermost URL instead of requiring a unique repository format,
which would improve intuitivity.  Additionally, the channel/team names
are easier for users to find than the channel ID, which means a smoother
user experience.

This does not break the existing workflow, and the previous usage still
exists, just with the ability to alternatively specify names.
Similarly, any new function parameters use keyword arguments, as to
prevent creating a breaking change.

Amended from add8a2b:  Appease flake8
Amended from c70a6a2:  Include `team` in origin when specified
Amended from 5baa610:  Create tests for the new functions/cases
Amended from 1f12371:  Remembered that tests need flake8 too
Amended from 881ee93:  Added a new test for the command parser

Signed-off-by: Emi Simpson <emi@alchemi.dev>